The most significant changes in the code include the reorganization of namespaces for better code organization, the removal and replacement of several repository interfaces and their corresponding classes, and the addition of new service classes. The namespaces of several interfaces and classes have been changed, for instance, `TheBigStore.Service.Interfaces` has been changed to `TheBigStore.Service.Interfaces.UserInterfaces` and `TheBigStore.Repository.Interfaces` has been changed to `TheBigStore.Repository.Interfaces.GenericInterfaces`.
Several repository interfaces and their corresponding classes have been removed and replaced with new versions in the `OrderInterfaces` and `UserInterfaces` namespaces. The `IGenericService` interface and its corresponding classes have been removed and replaced with a new version in the `GenericInterfaces` namespace. The `ICustomerService`, `IRoleService`, and `IUserService` interfaces have been moved to the `UserInterfaces` namespace.

New service classes have been added, including `GenericService`, `CustomerService`, `RoleService`, and `UserService`. These classes provide methods for creating, deleting, updating, and retrieving entities, as well as specific methods for the `Customer`, `Role`, and `User` entities.

Changes:

1. Namespaces of several interfaces and classes have been changed for better code organization.
2. The `Program.cs` file has been updated to use the new namespaces and to include additional services and repositories.
3. Several repository interfaces and their corresponding classes have been removed and replaced with new versions.
4. The `IGenericService` interface and its corresponding classes have been removed and replaced with a new version.
5. The `ICustomerService`, `IRoleService`, and `IUserService` interfaces have been moved to a new namespace.
6. The `MappingService` class has been removed.
7. The `GenericService` class has been removed and replaced with a new version.
8. The `CustomerService`, `RoleService`, and `UserService` classes have been removed.
9. The `GenericService.cs`, `MappingService.cs`, `CustomerService.cs`, `RoleService.cs`, and `UserService.cs` files have been added.
