+++ TheBigStore.Application/Pages/Roles/EditRoles.cshtml.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using TheBigStore.Repository.Models;
-using TheBigStore.Service.Interfaces;
+using TheBigStore.Service.Interfaces.UserInterfaces;
 
 namespace TheBigStore.Application.Pages.Roles
 {
