--- TheBigStore.Repository/Interfaces/IAddressRepository.cs
@@ -1,9 +1,0 @@
-﻿using TheBigStore.Repository.Interfaces;
-using TheBigStore.Repository.Models;
-
-namespace TheBigStore.Repository.Repositories
-{
-    public interface IAddressRepository : IGenericRepository<Address>
-    {
-    }
-}
