--- TheBigStore.Repository/Interfaces/ICustomerRepository.cs
@@ -1,8 +1,0 @@
-﻿using TheBigStore.Repository.Models;
-
-namespace TheBigStore.Repository.Interfaces
-{
-    public interface ICustomerRepository : IGenericRepository<Customer>
-    {
-    }
-}
