--- TheBigStore.Repository/Interfaces/IItemOrderRepository.cs
@@ -1,9 +1,0 @@
-﻿using TheBigStore.Repository.Interfaces;
-using TheBigStore.Repository.Models;
-
-namespace TheBigStore.Repository.Repositories
-{
-    public interface IItemOrderRepository : IGenericRepository<ItemOrder>
-    {
-    }
-}
