--- TheBigStore.Repository/Interfaces/IRoleRepository.cs
@@ -1,9 +1,0 @@
-﻿using TheBigStore.Repository.Interfaces;
-using TheBigStore.Repository.Models;
-
-namespace TheBigStore.Repository
-{
-    public interface IRoleRepository : IGenericRepository<Role>
-    { 
-    }
-}
