+++ TheBigStore.Repository/Interfaces/OrderInterfaces/IItemOrderRepository.cs
@@ -1,0 +1,9 @@
+﻿using TheBigStore.Repository.Interfaces.GenericInterfaces;
+using TheBigStore.Repository.Models;
+
+namespace TheBigStore.Repository.Interfaces.OrderInterfaces
+{
+    public interface IItemOrderRepository : IGenericRepository<ItemOrder>
+    {
+    }
+}
