+++ TheBigStore.Repository/Interfaces/OrderInterfaces/IItemRepository.cs
@@ -1,0 +1,9 @@
+﻿using TheBigStore.Repository.Interfaces.GenericInterfaces;
+using TheBigStore.Repository.Models;
+
+namespace TheBigStore.Repository.Interfaces.OrderInterfaces
+{
+    public interface IItemRepository : IGenericRepository<Item>
+    {
+    }
+}
