+++ TheBigStore.Repository/Interfaces/UserInterfaces/IAddressRepository.cs
@@ -1,0 +1,9 @@
+﻿using TheBigStore.Repository.Interfaces.GenericInterfaces;
+using TheBigStore.Repository.Models;
+
+namespace TheBigStore.Repository.Interfaces.UserInterfaces
+{
+    public interface IAddressRepository : IGenericRepository<Address>
+    {
+    }
+}
