+++ TheBigStore.Repository/Interfaces/UserInterfaces/ICustomerRepository.cs
@@ -1,0 +1,9 @@
+﻿using TheBigStore.Repository.Interfaces.GenericInterfaces;
+using TheBigStore.Repository.Models;
+
+namespace TheBigStore.Repository.Interfaces.UserInterfaces
+{
+    public interface ICustomerRepository : IGenericRepository<Customer>
+    {
+    }
+}
