+++ TheBigStore.Repository/Interfaces/UserInterfaces/IRoleRepository.cs
@@ -1,0 +1,9 @@
+﻿using TheBigStore.Repository.Interfaces.GenericInterfaces;
+using TheBigStore.Repository.Models;
+
+namespace TheBigStore.Repository.Interfaces.UserInterfaces
+{
+    public interface IRoleRepository : IGenericRepository<Role>
+    {
+    }
+}
