--- TheBigStore.Repository/Repositories/CategoryRepository.cs
@@ -1,9 +1,0 @@
-﻿using TheBigStore.Repository.Domain;
-using TheBigStore.Repository.Models;
-
-namespace TheBigStore.Repository.Repositories
-{
-    public class CategoryRepository(TheBigStoreContext dbContext) : GenericRepository<Category>(dbContext), ICategoryRepository
-    {
-    }
-}
