--- TheBigStore.Repository/Repositories/CustomerRepository.cs
@@ -1,10 +1,0 @@
-﻿using TheBigStore.Repository.Domain;
-using TheBigStore.Repository.Interfaces;
-using TheBigStore.Repository.Models;
-
-namespace TheBigStore.Repository.Repositories
-{
-    public class CustomerRepository(TheBigStoreContext dbContext) : GenericRepository<Customer>(dbContext), ICustomerRepository
-    {
-    }
-}
