+++ TheBigStore.Repository/Repositories/OrderRepositories/ItemOrderRepository.cs
@@ -1,0 +1,11 @@
+﻿using TheBigStore.Repository.Domain;
+using TheBigStore.Repository.Interfaces.OrderInterfaces;
+using TheBigStore.Repository.Models;
+using TheBigStore.Repository.Repositories.GenericRepositories;
+
+namespace TheBigStore.Repository.Repositories.OrderRepositories
+{
+    public class ItemOrderRepository(TheBigStoreContext dbContext) : GenericRepository<ItemOrder>(dbContext), IItemOrderRepository
+    {
+    }
+}
