--- TheBigStore.Repository/Repositories/OrderRepository.cs
@@ -1,9 +1,0 @@
-﻿using TheBigStore.Repository.Domain;
-using TheBigStore.Repository.Models;
-
-namespace TheBigStore.Repository.Repositories
-{
-    public class OrderRepository(TheBigStoreContext dbContext) : GenericRepository<Order>(dbContext), IOrderRepository
-    {
-    }
-}
