--- TheBigStore.Repository/Repositories/RoleRepository.cs
@@ -1,9 +1,0 @@
-﻿using TheBigStore.Repository.Domain;
-using TheBigStore.Repository.Models;
-
-namespace TheBigStore.Repository.Repositories
-{
-    public class RoleRepository(TheBigStoreContext dbContext) : GenericRepository<Role>(dbContext), IRoleRepository
-    {
-    }
-}
