--- TheBigStore.Service/Interfaces/IRoleService.cs
@@ -1,8 +1,0 @@
-﻿using TheBigStore.Repository.Models;
-
-namespace TheBigStore.Service.Interfaces
-{
-    public interface IRoleService : IGenericService<RoleDto>
-    {
-    }
-}
