+++ TheBigStore.Service/Interfaces/UserInterfaces/ICustomerService.cs
@@ -1,0 +1,9 @@
+﻿using TheBigStore.Repository.Models;
+using TheBigStore.Service.Interfaces.GenericInterfaces;
+
+namespace TheBigStore.Service.Interfaces.UserInterfaces
+{
+    public interface ICustomerService : IGenericService<CustomerDto>
+    {
+    }
+}
